+++ rustfst/src/algorithms/compose/lookahead_filters/push_weights_compose_filter.rs
@@ -1,6 +1,6 @@
 use std::cell::RefCell;
 use std::marker::PhantomData;
-use std::rc::Rc;
+use std::sync::Arc;
 
 use anyhow::Result;
 
@@ -38,9 +38,9 @@ where
     type M2 = CF::M2;
     type FS = PairFilterState<CF::FS, WeightFilterState<W>>;
 
-    fn new<IM1: Into<Option<Rc<RefCell<Self::M1>>>>, IM2: Into<Option<Rc<RefCell<Self::M2>>>>>(
-        fst1: Rc<<Self::M1 as Matcher<W>>::F>,
-        fst2: Rc<<Self::M2 as Matcher<W>>::F>,
+    fn new<IM1: Into<Option<Arc<RefCell<Self::M1>>>>, IM2: Into<Option<Arc<RefCell<Self::M2>>>>>(
+        fst1: Arc<<Self::M1 as Matcher<W>>::F>,
+        fst2: Arc<<Self::M2 as Matcher<W>>::F>,
         m1: IM1,
         m2: IM2,
     ) -> Result<Self> {
@@ -108,11 +108,11 @@ where
         w1.divide_assign(fweight, DivideType::DivideAny)
     }
 
-    fn matcher1(&self) -> Rc<RefCell<Self::M1>> {
+    fn matcher1(&self) -> Arc<RefCell<Self::M1>> {
         self.filter.matcher1()
     }
 
-    fn matcher2(&self) -> Rc<RefCell<Self::M2>> {
+    fn matcher2(&self) -> Arc<RefCell<Self::M2>> {
         self.filter.matcher2()
     }
 }