+++ rustfst/src/fst_impls/arc.rs
@@ -1,5 +1,5 @@
 use std::ops::Deref;
-use std::rc::Rc;
+use std::sync::Arc;
 
 use anyhow::Result;
 
@@ -8,36 +8,36 @@ use crate::fst_traits::{
 };
 use crate::SymbolTable;
 
-impl<F: Fst> Fst for Rc<F>
+impl<F: Fst> Fst for Arc<F>
 where
     F::W: 'static,
 {
-    fn input_symbols(&self) -> Option<Rc<SymbolTable>> {
+    fn input_symbols(&self) -> Option<&Arc<SymbolTable>> {
         self.deref().input_symbols()
     }
 
-    fn output_symbols(&self) -> Option<Rc<SymbolTable>> {
+    fn output_symbols(&self) -> Option<&Arc<SymbolTable>> {
         self.deref().output_symbols()
     }
 
-    fn set_input_symbols(&mut self, _symt: Rc<SymbolTable>) {
+    fn set_input_symbols(&mut self, _symt: Arc<SymbolTable>) {
         unimplemented!()
     }
 
-    fn set_output_symbols(&mut self, _symt: Rc<SymbolTable>) {
+    fn set_output_symbols(&mut self, _symt: Arc<SymbolTable>) {
         unimplemented!()
     }
 
-    fn unset_input_symbols(&mut self) -> Option<Rc<SymbolTable>> {
+    fn take_input_symbols(&mut self) -> Option<Arc<SymbolTable>> {
         unimplemented!()
     }
 
-    fn unset_output_symbols(&mut self) -> Option<Rc<SymbolTable>> {
+    fn take_output_symbols(&mut self) -> Option<Arc<SymbolTable>> {
         unimplemented!()
     }
 }
 
-impl<F: ExpandedFst> ExpandedFst for Rc<F>
+impl<F: ExpandedFst> ExpandedFst for Arc<F>
 where
     F::W: 'static,
 {
@@ -46,7 +46,7 @@ where
     }
 }
 
-impl<F: CoreFst> CoreFst for Rc<F> {
+impl<F: CoreFst> CoreFst for Arc<F> {
     type W = F::W;
 
     fn start(&self) -> Option<usize> {
@@ -70,7 +70,7 @@ impl<F: CoreFst> CoreFst for Rc<F> {
     }
 }
 
-impl<'a, F: FstIterator<'a>> FstIterator<'a> for Rc<F>
+impl<'a, F: FstIterator<'a>> FstIterator<'a> for Arc<F>
 where
     F::W: 'a,
 {
@@ -82,7 +82,7 @@ where
     }
 }
 
-impl<'a, F: TrIterator<'a>> TrIterator<'a> for Rc<F>
+impl<'a, F: TrIterator<'a>> TrIterator<'a> for Arc<F>
 where
     F::W: 'a,
 {
@@ -97,7 +97,7 @@ where
     }
 }
 
-impl<'a, F: StateIterator<'a>> StateIterator<'a> for Rc<F> {
+impl<'a, F: StateIterator<'a>> StateIterator<'a> for Arc<F> {
     type Iter = F::Iter;
 
     fn states_iter(&'a self) -> Self::Iter {
@@ -105,7 +105,7 @@ impl<'a, F: StateIterator<'a>> StateIterator<'a> for Rc<F> {
     }
 }
 
-impl<F: FstIntoIterator> FstIntoIterator for Rc<F> {
+impl<F: FstIntoIterator> FstIntoIterator for Arc<F> {
     type TrsIter = F::TrsIter;
     type FstIter = F::FstIter;