+++ rustfst/src/tests_openfst/fst_impls/const_fst.rs
@@ -1,4 +1,4 @@
-use std::rc::Rc;
+use std::sync::Arc;
 
 use anyhow::Result;
 
@@ -22,8 +22,8 @@ where
         osymt.add_symbol("b");
         osymt.add_symbol("c");
 
-        raw_fst.set_input_symbols(Rc::new(isymt));
-        raw_fst.set_output_symbols(Rc::new(osymt));
+        raw_fst.set_input_symbols(Arc::new(isymt));
+        raw_fst.set_output_symbols(Arc::new(osymt));
     }
 
     let const_fst: ConstFst<_> = raw_fst.clone().into();